pyproject.toml
[tool.poetry]
name = "BuyMeACoffee Fetcher"
version = "0.1.0"
description = "Fetch all donations associated with a BuyMeACoffee account and write output to `./donations.json`."
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
maintainers = ["Todd Birchard <toddbirchard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/toddbirchard/buymeacoffee-fetcher/"
repository = "https://github.com/toddbirchard/buymeacoffee-fetcher/"
keywords = [
  "BuyMeACoffee",
  "Requests",
  "Automation",
  "I/O"
]

[tool.poetry.dependencies]
python ="^3.9"
requests = "*"
loguru = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mypy = "*"
black = "*"
isort = "*"

[tool.poetry.scripts]
run = "main:fetch_donations"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/buymeacoffee-fetcher/issues"

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["buymeacoffee"]

[tool.black]
line-length = 80

[tool.pylint.'MESSAGES CONTROL']
disable="C0103,C0301,W0703,W0621"
